# docs/03-qa-review/dochub-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DocHub 后台管理接口",
    "version": "1.0.0",
    "description": "DocHub 企业文档管理平台 - RBAC 权限体系接口（用户/部门/角色/权限/授权绑定）。\n\n导入 Apifox 后直接使用；上下文路径为 /backend，见 servers[0].url。\n所有响应统一为 Result 包装体：{ code, message, data }。"
  },
  "servers": [
    { "url": "http://localhost:10086/backend", "description": "本地开发环境" }
  ],
...
  "paths": {
    "/sys/user": {
      "post": {
        "tags": ["用户管理"],
        "summary": "创建用户",
        "operationId": "createUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SysUserCreateDto" },
              "example": {
                "username": "zhangsan",
                "password": "Passw0rd!",
                "realName": "张三",
                "email": "zhangsan@example.com",
                "phone": "13800138000",
                "deptId": 1,
                "status": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResultSysUserVo" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["用户管理"],
        "summary": "更新用户（不含密码）",
        "operationId": "updateUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SysUserUpdateDto" },
              "example": {
                "id": 1,
                "realName": "张三丰",
                "email": "zhangsan@example.com",
                "phone": "13900139000",
                "deptId": 2,
                "status": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResultSysUserVo" }
              }
            }
          }
        }
      }
    },
    "/sys/user/{id}": {
      "get": {
        "tags": ["用户管理"],
        "summary": "查询单个用户",
        "operationId": "getUserById",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResultSysUserVo" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["用户管理"],
        "summary": "删除用户（逻辑删除）",
        "operationId": "deleteUser",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Result" }
              }
            }
          }
        }
      }
    },
    "/sys/user/list": {
      "get": {
        "tags": ["用户管理"],
        "summary": "分页查询用户",
        "operationId": "pageUser",
        "parameters": [
          { "name": "pageNum", "in": "query", "required": false, "schema": { "type": "integer", "default": 1 }, "example": 1, "description": "页码，从 1 开始" },
          { "name": "pageSize", "in": "query", "required": false, "schema": { "type": "integer", "default": 10 }, "example": 10, "description": "每页条数" },
          { "name": "deptId", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" }, "description": "按部门过滤" },
          { "name": "keyword", "in": "query", "required": false, "schema": { "type": "string" }, "description": "账号/姓名模糊匹配" }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResultPageSysUserVo" }
              }
            }
          }
        }
      }
    },
...
    "/sys/auth/user/{userId}": {
      "get": {
        "tags": ["授权绑定"],
        "summary": "查询用户绑定的角色与直挂权限",
        "operationId": "getUserAuth",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResultUserAuth" },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "roleIds": [1, 2],
                    "permissionIds": [1]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "ok" },
          "data": { "type": "object", "nullable": true }
        }
      },
      "SysUserCreateDto": {
        "type": "object",
        "required": ["username", "password", "deptId"],
        "properties": {
          "username": { "type": "string", "minLength": 3, "maxLength": 50, "example": "zhangsan", "description": "登录账号，唯一，3~50" },
          "password": { "type": "string", "minLength": 8, "maxLength": 64, "example": "Passw0rd!", "description": "密码，8~64" },
          "realName": { "type": "string", "maxLength": 50, "example": "张三" },
          "email": { "type": "string", "maxLength": 100, "example": "zhangsan@example.com" },
          "phone": { "type": "string", "example": "13800138000", "description": "1开头11位" },
          "deptId": { "type": "integer", "format": "int64", "example": 1, "description": "所属部门ID" },
          "status": { "type": "integer", "default": 0, "example": 0, "description": "0正常 1停用" }
        }
      },
      "SysUserUpdateDto": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "integer", "format": "int64", "example": 1 },
          "realName": { "type": "string", "maxLength": 50, "example": "张三丰" },
          "email": { "type": "string", "maxLength": 100, "example": "zhangsan@example.com" },
          "phone": { "type": "string", "example": "13900139000" },
          "deptId": { "type": "integer", "format": "int64", "example": 2 },
          "status": { "type": "integer", "example": 0 }
        }
      },
      "SysUserVo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "123456789012345" },
          "username": { "type": "string", "example": "zhangsan" },
          "realName": { "type": "string", "example": "张三" },
          "email": { "type": "string", "example": "zhangsan@example.com" },
          "phone": { "type": "string", "example": "13800138000" },
          "deptId": { "type": "string", "example": "1" },
          "status": { "type": "integer", "example": 0 },
          "createTime": { "type": "string", "format": "date-time", "example": "2026-09-20T10:00:00" },
          "updateTime": { "type": "string", "format": "date-time", "example": "2026-09-20T10:00:00" }
        }
      },
...
      "ResultUserAuth": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "ok" },
          "data": {
            "type": "object",
            "properties": {
              "roleIds": { "type": "array", "items": { "type": "integer", "format": "int64" }, "example": [1, 2] },
              "permissionIds": { "type": "array", "items": { "type": "integer", "format": "int64" }, "example": [1] }
            }
          }
        }
      },
...
      "PageResultBase": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "format": "int64", "example": 100 },
          "pageNum": { "type": "integer", "example": 1 },
          "pageSize": { "type": "integer", "example": 10 }
        }
      },
      "ResultSysUserVo": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "ok" },
          "data": { "$ref": "#/components/schemas/SysUserVo" }
        }
      },
...
      "ResultPageSysUserVo": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "message": { "type": "string", "example": "ok" },
          "data": {
            "allOf": [
              { "$ref": "#/components/schemas/PageResultBase" },
              {
                "type": "object",
                "properties": {
                  "records": { "type": "array", "items": { "$ref": "#/components/schemas/SysUserVo" } }
                }
              }
            ]
          }
        }
      },
...
    }
  }
}